chore: move NullableExtension.kt outside graphql package
