+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/graphql/supply/reservation/function/ConfirmReservationNotificationFun.kt
@@ -1,7 +1,7 @@
 package com.expediagroup.sdk.lodgingconnectivity.graphql.supply.reservation.function
 
 import com.expediagroup.sdk.lodgingconnectivity.graphql.common.GraphQLExecutor
-import com.expediagroup.sdk.lodgingconnectivity.graphql.extension.orFalseIfNull
+import com.expediagroup.sdk.lodgingconnectivity.extension.orFalseIfNull
 import com.expediagroup.sdk.lodgingconnectivity.graphql.model.response.RawResponse
 import com.expediagroup.sdk.lodgingconnectivity.graphql.model.response.Response
 import com.expediagroup.sdk.lodgingconnectivity.graphql.supply.ConfirmReservationNotificationMutation
