+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/graphql/supply/reservation/function/GetPropertyReservationsFun.kt
@@ -4,9 +4,9 @@ package com.expediagroup.sdk.lodgingconnectivity.graphql.supply.reservation.func
 
 import com.expediagroup.sdk.core.model.exception.service.ExpediaGroupServiceException
 import com.expediagroup.sdk.lodgingconnectivity.graphql.common.GraphQLExecutor
-import com.expediagroup.sdk.lodgingconnectivity.graphql.extension.getOrThrow
-import com.expediagroup.sdk.lodgingconnectivity.graphql.extension.orFalseIfNull
-import com.expediagroup.sdk.lodgingconnectivity.graphql.extension.orNullIfBlank
+import com.expediagroup.sdk.lodgingconnectivity.extension.getOrThrow
+import com.expediagroup.sdk.lodgingconnectivity.extension.orFalseIfNull
+import com.expediagroup.sdk.lodgingconnectivity.extension.orNullIfBlank
 import com.expediagroup.sdk.lodgingconnectivity.graphql.model.paging.PageInfo
 import com.expediagroup.sdk.lodgingconnectivity.graphql.model.response.PaginatedResponse
 import com.expediagroup.sdk.lodgingconnectivity.graphql.model.response.RawResponse
